+++ bidict/__init__.py
@@ -58,10 +58,10 @@ if _version_info < (3, 6):  # pragma: no cover
 
 # The rest of this file only collects functionality implemented in the rest of the
 # source for the purposes of exporting it under the `bidict` module namespace.
-# pylint: disable=wrong-import-position
 # flake8: noqa: F401 (imported but unused)
+from ._typing import KT, VT, DT, OKT, OVT, ODT, _NONE, IterItems, MapOrIterItems
 from ._abc import BidirectionalMapping, MutableBidirectionalMapping
-from ._base import BidictBase
+from ._base import BidictBase, BT
 from ._mut import MutableBidict
 from ._bidict import bidict
 from ._frozenbidict import frozenbidict
@@ -69,19 +69,23 @@ from ._frozenordered import FrozenOrderedBidict
 from ._named import namedbidict
 from ._orderedbase import OrderedBidictBase
 from ._orderedbidict import OrderedBidict
-from ._dup import (
-    ON_DUP_DEFAULT, ON_DUP_RAISE, ON_DUP_DROP_OLD,
-    RAISE, DROP_OLD, DROP_NEW, OnDup, OnDupAction,
-)
-from ._exc import (
-    BidictException,
-    DuplicationError, KeyDuplicationError, ValueDuplicationError, KeyAndValueDuplicationError,
-)
+from ._dup import ON_DUP_DEFAULT, ON_DUP_RAISE, ON_DUP_DROP_OLD, RAISE, DROP_OLD, DROP_NEW, OnDup, OnDupAction
+from ._exc import BidictException, DuplicationError, KeyDuplicationError, ValueDuplicationError, KeyAndValueDuplicationError
 from ._util import inverted
 from .metadata import (
     __author__, __maintainer__, __copyright__, __email__, __credits__, __url__,
     __license__, __status__, __description__, __keywords__, __version__, __version_info__,
 )
+
+# Set __module__ of re-exported classes to 'bidict' for nicer class repr strings:
+_locals = tuple(locals().items())
+for _name, _obj in _locals:  # pragma: no cover
+    if not getattr(_obj, '__module__', '').startswith('bidict.'):
+        continue
+    try:
+        _obj.__module__ = 'bidict'
+    except AttributeError as exc:  # raised when __module__ is read-only (as in OnDup)
+        pass
 
 
 #                             * Code review nav *