+++ bidict/_bidict.py
@@ -28,15 +28,21 @@
 
 """Provide :class:`bidict`."""
 
-from ._base import KT, VT
-from ._mut import MutableBidict
+import typing as _t
+
 from ._delegating import _DelegatingBidict
+from ._mut import MutableBidict
+from ._typing import KT, VT
 
 
 class bidict(_DelegatingBidict[KT, VT], MutableBidict[KT, VT]):
     """Base class for mutable bidirectional mappings."""
 
     __slots__ = ()
+
+    if _t.TYPE_CHECKING:  # pragma: no cover
+        @property
+        def inverse(self) -> 'bidict[VT, KT]': ...
 
 
 #                             * Code review nav *