+++ docs/conf.py
@@ -23,15 +23,12 @@ import sys
 import os
 
 
-# pylint: disable=invalid-name
-
-
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 sys.path.insert(0, os.path.abspath('..'))
 
-import bidict  # noqa: E402; pylint: disable=wrong-import-position
+import bidict
 
 # -- General configuration ------------------------------------------------
 
@@ -41,8 +38,7 @@ suppress_warnings = ['image.nonlocal_uri']
 #needs_sphinx = '1.0'
 
 # Add any Sphinx extension module names here, as strings. They can be
-# extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
-# ones.
+# extensions coming with Sphinx (named 'sphinx.ext.*') or custom ones.
 extensions = [
     'alabaster',
     'sphinx.ext.autodoc',
@@ -52,6 +48,7 @@ extensions = [
     'sphinx.ext.intersphinx',
     'sphinx.ext.viewcode',
     'sphinx.ext.todo',
+    'sphinx_autodoc_typehints',
 ]
 
 intersphinx_mapping = {'python': ('https://docs.python.org/3', None)}
@@ -72,7 +69,7 @@ master_doc = 'index'
 # General information about the project.
 project = 'bidict'
 author = bidict.__author__
-copyright = bidict.__copyright__  # pylint: disable=redefined-builtin
+copyright = bidict.__copyright__
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
@@ -144,7 +141,7 @@ html_theme_options = dict(
     show_powered_by=False,
     show_relbar_bottom=True,
     donate_url='https://gumroad.com/l/bidict',
-    tidelift_url='https://tidelift.com/subscription/pkg/pypi-bidict?utm_source=pypi-bidict&utm_medium=referral&utm_campaign=docs',  # noqa: E501; pylint: disable=line-too-long
+    tidelift_url='https://tidelift.com/subscription/pkg/pypi-bidict?utm_source=pypi-bidict&utm_medium=referral&utm_campaign=docs',
 )
 
 # Add any paths that contain custom themes here, relative to this directory.
@@ -309,6 +306,7 @@ htmlhelp_basename = 'bidictdoc'
 
 # Ignore urls matching these regex strings when doing "make linkcheck"
 linkcheck_ignore = [
+    r'http://ignore\.com/.*'
 ]
 linkcheck_timeout = 30  # 5s default too low
 
@@ -321,5 +319,5 @@ doctest_global_setup = """
 
 
 def setup(app):
-    """https://docs.readthedocs.io/en/latest/guides/adding-custom-css.html#adding-custom-css-or-javascript-to-a-sphinx-project"""  # noqa: E501; pylint: disable=line-too-long
-    app.add_javascript('custom.js')
+    """https://docs.readthedocs.io/en/latest/guides/adding-custom-css.html#adding-custom-css-or-javascript-to-a-sphinx-project"""
+    app.add_js_file('custom.js')